refactor(core): folder structure
