+++ packages/core/src/components/icon/icon.scss
@@ -1,4 +1,4 @@
-@use '../../core' as castor;
+@use '../../helpers' as castor;
 
 @mixin Icon() {
   $size: castor.space(3);
