--- packages/core/src/core/index.scss
@@ -1,5 +1,0 @@
-@forward './helpers';
-@forward './theme';
-@forward './themes/day';
-@forward './themes/night';
-@forward './tokens';
