--- packages/core/src/core/index.ts
@@ -1,2 +1,0 @@
-export * from './helpers';
-export { switchTheme } from './themes/helpers/switchTheme';
