+++ packages/core/src/helpers/index.ts
@@ -2,3 +2,4 @@ export * from './border-radius/border-radius';
 export * from './color/color';
 export * from './font/font';
 export * from './space/space';
+export * from './switch-theme/switchTheme';
