+++ packages/core/src/helpers/switch-theme/switchTheme.spec.ts
