+++ packages/core/src/helpers/switch-theme/switchTheme.ts
