+++ packages/core/src/index.scss
@@ -1,2 +1,3 @@
 @forward './components';
-@forward './core';
+@forward './helpers';
+@forward './themes';
