+++ packages/core/src/index.ts
@@ -1,3 +1,3 @@
 export * from './components';
-export * from './core';
+export * from './helpers';
 export * from './utils';
