+++ packages/core/src/themes/day.scss
@@ -1,4 +1,4 @@
-@use '../theme' as castor;
+@use './theme' as castor;
 
 // ...$types: 'class' | 'raw';
 @mixin day($types...) {
