+++ packages/core/src/themes/index.scss
@@ -1,0 +1,4 @@
+@forward './day';
+@forward './night';
+@forward './theme';
+@forward './tokens';
