+++ packages/core/src/themes/night.scss
@@ -1,4 +1,4 @@
-@use '../theme' as castor;
+@use './theme' as castor;
 
 // ...$types: 'class' | 'raw';
 @mixin night($types...) {
