+++ packages/core/src/utils/with-focus.scss
@@ -1,4 +1,4 @@
-@use '../core' as castor;
+@use '../helpers' as castor;
 
 $allowed-types: ('action', 'negative');
