+++ packages/core/style/themes/night-class.scss
@@ -1,4 +1,4 @@
 // this file will generate ./dist/themes/night-class.css
-@use '../../src/core/themes/night' as theme;
+@use '../../src/themes';
 
-@include theme.night('class');
+@include themes.night('class');
