+++ packages/core/style/themes/night.scss
@@ -1,6 +1,6 @@
 // this file will generate ./dist/themes/night.css
-@use '../../src/core/themes/night' as theme;
+@use '../../src/themes';
 
 :root {
-  @include theme.night();
+  @include themes.night();
 }
